1、采用加密方式初始化SDK，并配置设备列表 2、增加SDK初始化的文档说明

+++ app/src/main/java/com/lefu/ppblutoothkit/calculate/Calculate4ACActivitiy.kt
@@ -42,7 +42,7 @@ class Calculate4ACActivitiy : Activity() {
         val height = etHeight.text?.toString()?.toInt() ?: 180
         val age = etAge.text?.toString()?.toInt() ?: 28
         val weight = etWeight.text?.toString()?.toDouble() ?: 70.00
-        val impedance = etImpedance.text?.toString()?.toFloat() ?: 4195332L
+        val impedance = etImpedance.text?.toString()?.toLong() ?: 4195332L
 
         val userModel = PPUserModel.Builder()
             .setSex(sex) //gender
